+++ Creatures/IDamageable.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace DungeonExplorer
+{
+    interface IDamageable
+    {
+        int health;
+        int GetHealth();
+        void SetHealth(int healthChange);
+    }
+
+}
