+++ Creatures/IFightable.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace DungeonExplorer
+{
+    interface IFightable
+    {
+        void getAttackType();
+    }
+}
