+++ Items/IConsumable.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace DungeonExplorer
+{
+    interface IConsumable
+    {
+        int GetHpEffect();
+    }
+}
